фикс

+++ Content.Shared/Imperial/Medieval/Ships/Repairing/ShipRepairSystem.cs
@@ -91,7 +91,7 @@ public sealed class ShipRepairSystem : EntitySystem
         _stack.Use(uid, 1);
     }
 
-    private bool IsRepairableTile(ushort tileId)
+    private bool IsRepairableTile(int tileId)
     {
         var ids = new[]
         {
